Renamed search-groups to search and updated GroupItem and module loading names
