+++ src/data/index.js
@@ -51,8 +51,7 @@ import * as PhoneNumber from './phone-number';
 import * as PrayerRequest from './prayer-request';
 import * as RockConstants from './rock-constants';
 import * as Schedule from './schedule';
-// import * as Search from './search';
-import * as SearchGroups from './search-groups';
+import * as Search from './search';
 import * as StreamChat from './stream-chat';
 import * as TwilioNotify from './twilio-notify';
 import * as Url from './url';
@@ -105,8 +104,7 @@ const data = {
   RockConstants,
   Schedule,
   Scripture,
-  // Search,
-  SearchGroups,
+  Search,
   Sharable,
   Sms: TwilioNotify,
   StreamChat,
