+++ src/data/search/schema.js
@@ -1,19 +1,18 @@
 import gql from 'graphql-tag';
 
 export default gql`
-
   type SearchResultsConnection {
     edges: [SearchResult]
     pageInfo: PaginationInfo
   }
 
-  enum SearchIndexAction {
-    update
-    delete
-  }
-
   interface SearchResult {
     cursor: String
     node: Node
+  }
+
+  enum SearchIndexAction {
+    update
+    delete
   }
 `
